# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Easy-Stream API
  description: 直播流管理系统 API
  version: "2.0"
  contact:
    name: Easy-Stream
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发服务器
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Access Token
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误信息
    Message:
      type: object
      properties:
        message:
          type: string
          description: 消息
    HookResponse:
      type: object
      properties:
        code:
          type: integer
          description: 响应码，0 表示成功，-1 表示失败
        msg:
          type: string
          description: 响应消息
...
    RecordFile:
      type: object
      description: 录制文件信息
      properties:
        file_name:
          type: string
          description: 文件名（相对路径）
          example: "2024-01-15/12-30-45.mp4"
        file_path:
          type: string
          description: 完整文件路径
          example: "live/abc123def456/2024-01-15/12-30-45.mp4"
        file_size:
          type: integer
          format: int64
          description: 文件大小（字节）
          example: 104857600
        duration:
          type: number
          format: float
          description: 录制时长（秒）
          example: 3600.5
        start_time:
          type: integer
          format: int64
          description: 录制开始时间戳
          example: 1705315845
        time_len:
          type: number
          format: float
          description: 录制时长（秒，ZLM 返回）
          example: 3600.5
        created_at:
          type: string
          format: date-time
          description: 文件创建时间
        urls:
          type: object
          description: 各存储的访问URL
          additionalProperties:
            type: string
          example:
            download: "/api/v1/records/abc123def456/2024-01-15/12-30-45.mp4/download"
            s3: "https://bucket.s3.amazonaws.com/records/abc123def456/2024-01-15/12-30-45.mp4"
            cos: "https://bucket.cos.ap-guangzhou.myqcloud.com/records/abc123def456/2024-01-15/12-30-45.mp4"
paths:
...
  /hooks/on_publish:
    post:
      tags:
        - Hook
      summary: 推流开始回调
      description: |
        ZLMediaKit 推流开始时调用此接口进行验证。

        **验证逻辑：**
        - stream_key 必须存在
        - stream 状态不能为 ended

        验证失败返回 code=-1，ZLMediaKit 会自动断开推流连接。
      operationId: onPublish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                app:
                  type: string
                  description: 应用名
                  example: live
                stream:
                  type: string
                  description: 流名称（stream_key）
                schema:
                  type: string
                  description: 协议
                  example: rtmp
                mediaServerId:
                  type: string
                  description: 媒体服务器 ID
                ip:
                  type: string
                  description: 推流端 IP
                port:
                  type: integer
                  description: 推流端端口
      responses:
        '200':
          description: 响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HookResponse'
              examples:
                success:
                  summary: 验证通过
                  value:
                    code: 0
                    msg: success
                failed:
                  summary: 验证失败
                  value:
                    code: -1
                    msg: stream not found
  /hooks/on_unpublish:
    post:
      tags:
        - Hook
      summary: 推流结束回调
      description: ZLMediaKit 推流结束时调用此接口
      operationId: onUnpublish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                app:
                  type: string
                  description: 应用名
                stream:
                  type: string
                  description: 流名称
                schema:
                  type: string
                  description: 协议
                mediaServerId:
                  type: string
                  description: 媒体服务器 ID
      responses:
        '200':
          description: 响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HookResponse'
  /hooks/on_flow_report:
    post:
      tags:
        - Hook
      summary: 流量统计回调
      description: ZLMediaKit 流量统计回调
      operationId: onFlowReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                app:
                  type: string
                stream:
                  type: string
                totalBytes:
                  type: integer
                  description: 总字节数
                duration:
                  type: integer
                  description: 持续时间（秒）
                player:
                  type: boolean
                  description: 是否为播放器
      responses:
        '200':
          description: 响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HookResponse'
  /hooks/on_stream_none_reader:
    post:
      tags:
        - Hook
      summary: 无人观看回调
      description: ZLMediaKit 无人观看时调用此接口
      operationId: onStreamNoneReader
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                app:
                  type: string
                stream:
                  type: string
                schema:
                  type: string
                mediaServerId:
                  type: string
      responses:
        '200':
          description: 响应
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  close:
                    type: boolean
                    description: 是否关闭流

  # ==================== 录制文件管理接口 ====================
  /records:
    get:
      tags:
        - 录制文件
      summary: 获取所有录制文件列表
      description: 管理员获取所有直播的录制文件列表（分页）
      operationId: listAllRecords
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    description: 总记录数
                  streams:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        stream_key:
                          type: string
                        name:
                          type: string
                        record_status:
                          type: string
                        record_files:
                          type: array
                          items:
                            $ref: '#/components/schemas/RecordFile'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /records/{key}:
    get:
      tags:
        - 录制文件
      summary: 获取指定直播的录制文件
      description: 管理员获取指定直播的所有录制文件
      operationId: getRecordsByStreamKey
      security:
        - BearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          description: 推流密钥
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  stream_id:
                    type: integer
                  stream_key:
                    type: string
                  stream_name:
                    type: string
                  record_files:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecordFile'
                  record_status:
                    type: string
        '404':
          description: 直播不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /records/{key}/{filename}:
    delete:
      tags:
        - 录制文件
      summary: 删除录制文件
      description: |
        管理员删除指定的录制文件

        删除逻辑：
        1. 删除 ZLM 原始文件（本地模式）或标记删除（远程模式）
        2. 删除所有对象存储备份（S3/COS/OSS等）
        3. 删除数据库记录

        注意：远程模式下无法删除 ZLM 服务器上的物理文件，需要手动清理
      operationId: deleteRecordFile
      security:
        - BearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          description: 推流密钥
          schema:
            type: string
        - name: filename
          in: path
          required: true
          description: 文件名（需要 URL 编码）
          schema:
            type: string
          example: "2024-01-15%2F12-30-45.mp4"
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: 直播或文件不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 删除失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...